Inkludere "OKE_DELTAKELSE" i kodeverket, for bakoverkompatibilitet med Arena
