+++ src/main/java/no/nav/veilarbvedtaksstotte/domain/vedtak/HovedmalDetaljert.java
@@ -5,15 +5,16 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
 public enum HovedmalDetaljert {
-    SKAFFE_ARBEID(Hovedmal.SKAFFE_ARBEID, "Skaffe arbeid"),
-    BEHOLDE_ARBEID(Hovedmal.BEHOLDE_ARBEID, "Beholde arbeid");
+    SKAFFE_ARBEID(HovedmalMedOkeDeltakelse.SKAFFE_ARBEID, "Skaffe arbeid"),
+    BEHOLDE_ARBEID(HovedmalMedOkeDeltakelse.BEHOLDE_ARBEID, "Beholde arbeid"),
+    OKE_DELTAKELSE(HovedmalMedOkeDeltakelse.OKE_DELTAKELSE, "Øke deltakelse");
 
     @JsonProperty("kode")
-    Hovedmal kode;
+    HovedmalMedOkeDeltakelse kode;
     @JsonProperty("beskrivelse")
     String beskrivelse;
 
-    HovedmalDetaljert(Hovedmal kode, String beskrivelse) {
+    HovedmalDetaljert(HovedmalMedOkeDeltakelse kode, String beskrivelse) {
         this.kode = kode;
         this.beskrivelse = beskrivelse;
     }
